backend/requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
django-redis
